Use only device name/version for headset id
Otherwise different IPD setting would mean new device,
but with the same IP, it wouldn't show separately in the UI.
Also set IPD as a float field.

+++ alvr/client_hmd/app/src/main/cpp/ovr_context.cpp
@@ -1027,7 +1027,7 @@ void OvrContext::getDeviceDescriptor(JNIEnv *env, jobject deviceDescriptor) {
     fieldID = env->GetFieldID(clazz, "mControllerCapabilityFlags", "I");
     env->SetIntField(deviceDescriptor, fieldID, controllerCapabilityFlags);
     fieldID = env->GetFieldID(clazz, "mIpd", "F");
-    env->SetIntField(deviceDescriptor, fieldID, getIPD());
+    env->SetFloatField(deviceDescriptor, fieldID, getIPD());
 
     env->DeleteLocalRef(clazz);
 }
